Move executor into src/utils/
It's a generic subprocess/command utility (findBash, executeCommand,
executeArgv, killSession) — nothing ash-specific. Belonged at the top
level of src/ for historical reasons; the post-refactor tree has no
other top-level files, so it stands out. Moving it under utils/ leaves
src/ with only the five subsystem directories (core, shell, agent,
extensions, cli, utils).

Public 'agent-sh/executor' path preserved via updated exports map
pointing at dist/utils/executor.js. Consumers (ash tools, agent-loop,
superash) unaffected.
