+++ src/agent/agent-loop.ts
@@ -39,7 +39,7 @@ import { createToolProtocol, type ToolProtocol, type PendingToolCall as Protocol
 // Core tool factories
 import { createBashTool } from "./tools/bash.js";
 import { createPwshTool } from "./tools/pwsh.js";
-import { findBash } from "../executor.js";
+import { findBash } from "../utils/executor.js";
 import { createReadFileTool, type FileReadCache } from "./tools/read-file.js";
 import { createWriteFileTool } from "./tools/write-file.js";
 import { createEditFileTool } from "./tools/edit-file.js";
