+++ src/agent/tools/pwsh.ts
@@ -1,5 +1,5 @@
 import { spawnSync } from "node:child_process";
-import { executeArgv, killSession } from "../../executor.js";
+import { executeArgv, killSession } from "../../utils/executor.js";
 import type { EventBus } from "../../core/event-bus.js";
 import type { ToolDefinition } from "../types.js";
